+++ fpsi/fpsi_sender.h
@@ -100,7 +100,7 @@ public:
   void merge_timer(simpleTimer &other) { senderTimer.merge(other); }
 
   // 通信计数
-  std::vector<std::pair<string, u64>> commus;
+  std::vector<std::pair<string, double>> commus;
   void print_commus() {
     for (auto &x : commus) {
       spdlog::info("{}: {} MB", x.first, x.second);
